refactor(gax): move `ClientConfig` to `options`
`ClientConfig` is part of the public API, it needs documentation and it
needs to be available to use in the documentation of other features.

Moved it to a module that is **not** `doc(hidden)` and is **not**
conditional on unstable features.

This means the `auth` dependency is always required.
