+++ src/gax/src/http_client/mod.rs
@@ -94,135 +94,4 @@ impl std::fmt::Debug for ReqwestClient {
 #[derive(serde::Serialize)]
 pub struct NoBody {}
 
-#[derive(Default)]
-pub struct ClientConfig {
-    pub(crate) endpoint: Option<String>,
-    pub(crate) cred: Option<Credential>,
-    pub(crate) tracing: bool,
-}
-
-const LOGGING_VAR: &str = "GOOGLE_CLOUD_RUST_LOGGING";
-
-impl ClientConfig {
-    /// Returns a default [ConfigBuilder].
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    pub fn tracing_enabled(&self) -> bool {
-        if self.tracing {
-            return true;
-        }
-        std::env::var(LOGGING_VAR)
-            .map(|v| v == "true")
-            .unwrap_or(false)
-    }
-
-    /// Sets an endpoint that overrides the default endpoint for a service.
-    pub fn set_endpoint<T: Into<String>>(mut self, v: T) -> Self {
-        self.endpoint = Some(v.into());
-        self
-    }
-
-    /// Enables tracing.
-    pub fn enable_tracing(mut self) -> Self {
-        self.tracing = true;
-        self
-    }
-
-    /// Disables tracing.
-    pub fn disable_tracing(mut self) -> Self {
-        self.tracing = false;
-        self
-    }
-
-    pub fn set_credential<T: Into<Option<Credential>>>(mut self, v: T) -> Self {
-        self.cred = v.into();
-        self
-    }
-
-    pub(crate) async fn default_credential() -> Result<Credential> {
-        let cc = auth::CredentialConfig::builder()
-            .scopes(vec![
-                "https://www.googleapis.com/auth/cloud-platform".to_string()
-            ])
-            .build()
-            .map_err(Error::authentication)?;
-        Credential::find_default(cc)
-            .await
-            .map_err(Error::authentication)
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    type Result = std::result::Result<(), Box<dyn std::error::Error>>;
-
-    // This test must run serially because `std::env::remove_var` and
-    // `std::env::set_var` are unsafe otherwise.
-    #[test]
-    #[serial_test::serial]
-    fn config_tracing() {
-        unsafe {
-            std::env::remove_var(LOGGING_VAR);
-        }
-        let config = ClientConfig::new();
-        assert!(!config.tracing_enabled(), "expected tracing to be disabled");
-        let config = ClientConfig::new().enable_tracing();
-        assert!(config.tracing_enabled(), "expected tracing to be enabled");
-        let config = config.disable_tracing();
-        assert!(
-            !config.tracing_enabled(),
-            "expected tracing to be disaabled"
-        );
-
-        unsafe {
-            std::env::set_var(LOGGING_VAR, "true");
-        }
-        let config = ClientConfig::new();
-        assert!(config.tracing_enabled(), "expected tracing to be enabled");
-
-        unsafe {
-            std::env::set_var(LOGGING_VAR, "not-true");
-        }
-        let config = ClientConfig::new();
-        assert!(!config.tracing_enabled(), "expected tracing to be disabled");
-    }
-
-    #[test]
-    fn config_endpoint() {
-        let config = ClientConfig::new().set_endpoint("http://storage.googleapis.com");
-        assert_eq!(
-            config.endpoint,
-            Some("http://storage.googleapis.com".to_string())
-        );
-    }
-
-    #[tokio::test]
-    async fn config_credentials() -> Result {
-        let config = ClientConfig::new().set_credential(auth::Credential::test_credentials());
-        let cred = config.cred.unwrap();
-        let token = cred.access_token().await?;
-        assert!(
-            token.value.contains("test-only"),
-            "unexpected test token {}",
-            token.value
-        );
-        Ok(())
-    }
-
-    #[tokio::test]
-    #[serial_test::serial]
-    async fn config_default_credentials() -> Result {
-        let dir = tempfile::tempdir()?;
-        let path = dir.path().to_str().unwrap();
-        unsafe {
-            // This is not readable as a file and should cause the default credentials to fail.
-            std::env::set_var("GOOGLE_APPLICATION_CREDENTIALS", path);
-        }
-        let cred = ClientConfig::default_credential().await;
-        assert!(cred.is_err());
-        Ok(())
-    }
-}
+pub type ClientConfig = crate::options::ClientConfig;
